Add tests for onecall

+++ src/main/kotlin/org/openweather/cached/service/model/interfaces/Precipitation.kt
@@ -4,5 +4,5 @@ interface Precipitation {
     /**
      * Precipitation volume, mm
      */
-    val precipitation: Int
+    val precipitation: Float
 }
