+++ src/main/kotlin/org/openweather/cached/service/model/interfaces/PrecipitationLastHour.kt
@@ -4,5 +4,5 @@ interface PrecipitationLastHour {
     /**
      * Precipitation volume for last hour, mm
      */
-    val lastHour: Int?
+    val lastHour: Float?
 }
