+++ src/main/kotlin/org/openweather/cached/service/model/response/MinutelyWeather.kt
@@ -10,5 +10,5 @@ data class MinutelyWeather(
     override val dateTime: ZonedDateTime,
 
     @SerializedName("precipitation")
-    override val precipitation: Int,
+    override val precipitation: Float,
 ) : DateTime, Precipitation
