+++ src/main/kotlin/org/openweather/cached/service/model/response/Rain.kt
@@ -5,5 +5,5 @@ import org.openweather.cached.service.model.interfaces.PrecipitationLastHour
 
 data class Rain(
     @SerializedName("1h")
-    override val lastHour: Int?
+    override val lastHour: Float?
 ) : PrecipitationLastHour
